extract handleDoubleEqual and handleNestedParens
