+++ src/cleanupEquation.js
@@ -1,13 +1,10 @@
 import closeParens from "./closeParens";
 import addMultiplier from "./addMultiplier";
-import { evaluate, round } from "mathjs";
+import handleNestedParens from "./handleNestedParens";
+import handleDoubleEqual from "./handleDoubleEqual";
 
 const getLastChar = (from) => {
   return from.slice(-1);
-};
-const roundedResult = (expression) => {
-  const digitAfterComma = 13;
-  round(evaluate(expression), digitAfterComma);
 };
 
 const cleanupEquation = (result, operate, originalLastNum) => {
@@ -20,23 +17,8 @@ const cleanupEquation = (result, operate, originalLastNum) => {
 
   tempResult = closeParens(tempResult) + "";
   tempResult = addMultiplier(tempResult) + "";
-
-  // Handling double equal
-  // If I only have sign + digit(s) + dot + digit(s)
-  const regex = /-{0,1}[0123456789]*(\.[0123456789]*){0,1}/g;
-  const matches = tempResult.match(regex) || [];
-  if (matches.length > 1 && matches[0] === tempResult) {
-    //I need to handle the double equal
-    if ("+-/*".includes(operate) && !isNaN(originalLastNum)) {
-      // finalResult = evaluate(result + operate + originalLastNum) + "";
-      try {
-        finalResult =
-          roundedResult(tempResult + operate + originalLastNum) + "";
-      } catch (error) {
-        //throw error
-      }
-    }
-  }
+  finalResult = handleDoubleEqual(result, operate, originalLastNum) + "";
+  finalResult = handleNestedParens(tempResult) + "";
 
   return [tempResult, finalResult];
 };
